use existing navigatorexclude prop
